Move code related to sockets into its own file
