+++ src/index.js
@@ -3,8 +3,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import openSocket from 'socket.io-client';
+import getSocket from './Sockets/';
+//import openSocket from 'socket.io-client';
 import { processNewScores, processNewMessages } from './Actions/index';
+import type { Store } from 'redux';
 import store from './store';
 
 import App from './App';
@@ -18,14 +20,4 @@ ReactDOM.render(
 );
 registerServiceWorker();
 
- const socket = openSocket('http://localhost:4000');
-//const socket = openSocket('https://databraid.localtunnel.me');
-
-
-socket.on('messages', (messages) => {
-  store.dispatch(processNewMessages(messages));
-});
-
-socket.on('score', (scoreData) => {
-  store.dispatch(processNewScores(scoreData));
-});
+getSocket();
